chore: remove base Wallet component and move uti ls to DatePicker file
