--- packages/account-v2/src/components/base/WalletDropdown/WalletDropdown.scss
@@ -1,146 +1,0 @@
-/* stylelint-disable color-no-hex */
-
-.wallets-dropdown {
-    width: 100%;
-    position: relative;
-    cursor: pointer;
-
-    &--disabled {
-        pointer-events: none;
-
-        & label {
-            color: var(--system-light-5-active-background, #999);
-        }
-    }
-
-    &__button {
-        all: unset;
-        right: 1.6rem;
-        transform: rotate(0);
-        transform-origin: 50% 45%;
-        transition: transform 0.2s cubic-bezier(0.25, 0.1, 0.25, 1);
-
-        &--active {
-            transform: rotate(180deg);
-        }
-    }
-
-    &__content {
-        width: 100%;
-        background: var(--system-light-8-primary-background, #fff);
-        display: flex;
-        align-items: center;
-
-        .wallets-textfield__field {
-            cursor: pointer;
-        }
-    }
-
-    &__field {
-        position: absolute;
-        inset: 0;
-        min-width: 0; /* this is required to reset input's default width */
-        padding-left: 2rem;
-        display: flex;
-        flex-grow: 1;
-        font-family: inherit;
-        outline: 0;
-        font-size: 1.4rem;
-        background-color: transparent;
-        color: var(--system-light-2-general-text, #333);
-        transition: border-color 0.2s;
-        cursor: unset;
-        user-select: none;
-        &::selection {
-            background-color: transparent;
-        }
-
-        &::placeholder {
-            color: transparent;
-        }
-    }
-
-    &__field:placeholder-shown ~ &__label {
-        font-size: 1.4rem;
-        cursor: text;
-        top: 30%;
-        padding: 0;
-    }
-
-    &__field:placeholder-shown ~ &__label--with-icon {
-        left: 4.4rem;
-    }
-
-    label,
-    &__field:focus ~ &__label {
-        position: absolute;
-        top: 0;
-        display: block;
-        transition: 0.2s;
-        font-size: 1rem;
-        color: var(--system-light-3-less-prominent-text, #999);
-        background: var(--system-light-8-primary-background, #fff);
-        padding-inline: 0.4rem;
-        left: 1.6rem;
-    }
-
-    &__field:focus ~ &__label {
-        color: var(--brand-blue, #85acb0);
-    }
-
-    &__items {
-        position: absolute;
-        top: 4.8rem;
-        width: 100%;
-        display: flex;
-        flex-direction: column;
-        align-items: flex-start;
-        z-index: 2;
-        border-radius: 0.4rem;
-        background: var(--system-light-8-primary-background, #fff);
-        box-shadow: 0 3.2rem 6.4rem 0 rgba(14, 14, 14, 0.14);
-        overflow-y: auto;
-
-        & > :first-child {
-            border-radius: 0.4rem 0.4rem 0 0;
-        }
-
-        & > :last-child {
-            border-radius: 0 0 0.4rem 0.4rem;
-        }
-
-        &--sm {
-            max-height: 22rem;
-        }
-
-        &--md {
-            max-height: 42rem;
-        }
-
-        &--lg {
-            max-height: 66rem;
-        }
-    }
-
-    &__icon {
-        position: absolute;
-        left: 1.6rem;
-        width: 1.6rem;
-        height: 1.6rem;
-    }
-
-    &__item {
-        padding: 10px 16px;
-        width: 100%;
-        z-index: 2;
-
-        &:hover:not(&--active) {
-            cursor: pointer;
-            background: var(--system-light-6-hover-background, #e6e9e9);
-        }
-
-        &--active {
-            background: var(--system-light-5-active-background, #d6dadb);
-        }
-    }
-}
