--- packages/account-v2/src/components/base/WalletDropdown/index.ts
@@ -1,1 +1,0 @@
-export { default as WalletDropdown } from './WalletDropdown';
