--- packages/account-v2/src/components/base/WalletText/WalletText.scss
@@ -1,186 +1,0 @@
-/* stylelint-disable color-no-hex */
-
-$color-map: (
-    general: #333333,
-    primary: #999999,
-    success: #4bb4b3,
-    warning: #ffad3a,
-    error: #ec3f3f,
-    white: #ffffff,
-    black: #000000,
-    less-prominent: #999999,
-    red: #ff444f,
-    blue: #377cfc,
-    green: #17eabd,
-    orange: #ff9c13,
-    system-dark-2-general-text: #c2c2c2,
-);
-
-$desktop-font-size-map: (
-    '2xs': (
-        size: 1rem,
-        line_height: 1.4rem,
-    ),
-    'xs': (
-        size: 1.2rem,
-        line_height: 1.8rem,
-    ),
-    'sm': (
-        size: 1.4rem,
-        line_height: 2rem,
-    ),
-    'md': (
-        size: 1.6rem,
-        line_height: 2.4rem,
-    ),
-    'lg': (
-        size: 2rem,
-        line_height: 3rem,
-    ),
-    'xl': (
-        size: 2.4rem,
-        line_height: 3.6rem,
-    ),
-    '2xl': (
-        size: 3.2rem,
-        line_height: 4rem,
-    ),
-    '3xl': (
-        size: 4.8rem,
-        line_height: 6rem,
-    ),
-    '4xl': (
-        size: 6.4rem,
-        line_height: 8rem,
-    ),
-    '5xl': (
-        size: 8rem,
-        line_height: 10rem,
-    ),
-);
-
-$mobile-font-size-map: (
-    '2xs': (
-        size: 0.8rem,
-        line_height: 1.2rem,
-    ),
-    'xs': (
-        size: 1rem,
-        line_height: 1.4rem,
-    ),
-    'sm': (
-        size: 1.2rem,
-        line_height: 1.8rem,
-    ),
-    'md': (
-        size: 1.4rem,
-        line_height: 2rem,
-    ),
-    'lg': (
-        size: 1.6rem,
-        line_height: 2.4rem,
-    ),
-    'xl': (
-        size: 1.8rem,
-        line_height: 2.6rem,
-    ),
-    '2xl': (
-        size: 2.4rem,
-        line_height: 3rem,
-    ),
-    '3xl': (
-        size: 2.8rem,
-        line_height: 3.4rem,
-    ),
-    '4xl': (
-        size: 3.2rem,
-        line_height: 4rem,
-    ),
-    '5xl': (
-        size: 4rem,
-        line_height: 5rem,
-    ),
-);
-
-$line-height-map: (
-    '3xs': 1.2rem,
-    '2xs': 1.4rem,
-    'xs': 1.6rem,
-    'sm': 1.8rem,
-    'md': 2rem,
-    'lg': 2.2rem,
-    'xl': 2.4rem,
-    '2xl': 2.6rem,
-    '3xl': 2.8rem,
-    '4xl': 3rem,
-    '5xl': 3.2rem,
-    '6xl': 3.4rem,
-    '7xl': 3.6rem,
-);
-
-$font-weight-map: (
-    light: 300,
-    normal: 400,
-    bold: 700,
-);
-
-$font-align-map: (
-    left: left,
-    center: center,
-    right: right,
-);
-
-$font-style-map: (
-    italic: italic,
-    normal: normal,
-);
-
-.wallets-text {
-    @each $color, $value in $color-map {
-        &__color--#{$color} {
-            color: $value;
-        }
-    }
-
-    @each $size, $values in $desktop-font-size-map {
-        &__size--#{$size} {
-            @include desktop {
-                font-size: map-get($values, size);
-                line-height: map-get($values, line_height);
-            }
-        }
-    }
-
-    @each $size, $values in $mobile-font-size-map {
-        &__size--#{$size} {
-            @include mobile {
-                font-size: map-get($values, size);
-                line-height: map-get($values, line_height);
-            }
-        }
-    }
-
-    @each $lineHeight, $value in $line-height-map {
-        &__line-height--#{$lineHeight} {
-            line-height: $value;
-        }
-    }
-
-    @each $weight, $value in $font-weight-map {
-        &__weight--#{$weight} {
-            font-weight: $value;
-        }
-    }
-
-    @each $align, $value in $font-align-map {
-        &__align--#{$align} {
-            text-align: $value;
-        }
-    }
-
-    @each $fontStyle, $value in $font-style-map {
-        &__font-style--#{$fontStyle} {
-            font-style: $value;
-        }
-    }
-}
