--- packages/account-v2/src/components/base/WalletText/index.ts
@@ -1,1 +1,0 @@
-export { default as WalletText } from './WalletText';
