--- packages/account-v2/src/components/base/WalletTextField/WalletTextField.scss
@@ -1,145 +1,0 @@
-/* stylelint-disable color-no-hex */
-.wallets-textfield {
-    min-width: 12rem;
-    width: 100%;
-    position: relative;
-    display: flex;
-    flex-direction: column;
-    gap: 0.2rem;
-
-    &--error {
-        .wallets-textfield__box,
-        .wallets-textfield__box:hover {
-            border: 1px solid var(--status-light-danger, #ec3f3f);
-        }
-
-        .wallets-textfield__box:has(.wallets-textfield__field:focus) {
-            border: 1px solid var(--brand-blue, #ec3f3f);
-        }
-
-        .wallets-textfield__box:has(.wallets-textfield__field:valid) {
-            border: 1px solid var(--brand-blue, #ec3f3f);
-        }
-
-        .wallets-textfield__label {
-            color: #ec3f3f;
-        }
-
-        .wallets-textfield__field:focus ~ .wallets-textfield__label {
-            color: #ec3f3f;
-        }
-    }
-
-    &--disabled {
-        pointer-events: none;
-
-        & .wallets-textfield__box,
-        .wallets-textfield__box:hover {
-            border: 1px solid var(--system-light-5-active-background, #eaeced);
-
-            & input {
-                color: var(--system-light-5-active-background, #999);
-            }
-        }
-        & .wallets-textfield__box:has(.wallets-textfield__field:focus) {
-            border: 1px solid var(--system-light-5-active-background, #eaeced);
-        }
-        & .wallets-textfield__field {
-            background: inherit;
-        }
-    }
-
-    &__box {
-        height: 4rem;
-        width: 100%;
-        border-radius: 0.4rem;
-        padding: 1rem 1.6rem;
-        border: 1px solid var(--system-light-5-active-background, #d6dadb);
-        display: inline-flex;
-        align-items: center;
-        transition: border-color 0.2s;
-
-        &:hover {
-            border-color: var(--system-light-3-less-prominent-text, #999);
-        }
-    }
-
-    &__box:has(&__field:focus) {
-        border: 1px solid var(--brand-blue, #85acb0);
-    }
-
-    &__box:has(&__field:invalid) {
-        border: 1px solid var(--status-light-danger, #ec3f3f);
-    }
-
-    &__field {
-        min-width: 0;
-        font-family: inherit;
-        outline: 0;
-        font-size: 1.4rem;
-        color: var(--system-light-2-general-text, #333);
-        transition: border-color 0.2s;
-        flex: 1;
-    }
-
-    &__field::placeholder {
-        color: transparent;
-    }
-
-    &__field:placeholder-shown ~ &__label {
-        font-size: 1.4rem;
-        cursor: text;
-        top: 2rem;
-        padding: 0;
-    }
-
-    label,
-    &__field:focus ~ &__label {
-        position: absolute;
-        top: 0%;
-        transform: translateY(-50%);
-        display: block;
-        transition: 0.2s;
-        font-size: 1rem;
-        color: var(--system-light-3-less-prominent-text, #999);
-        background: var(--system-light-8-primary-background, #fff);
-        padding-inline: 0.4rem;
-        left: 1.6rem;
-        -webkit-user-select: none;
-        -ms-user-select: none;
-        user-select: none;
-    }
-
-    &__field:focus ~ &__label {
-        color: var(--brand-blue, #85acb0);
-    }
-
-    &__field:invalid ~ &__label {
-        color: var(--status-light-danger, #ec3f3f);
-    }
-
-    &__icon {
-        &-left {
-            margin-right: 0.8rem;
-        }
-
-        &-right {
-            margin-left: 1.6rem;
-        }
-    }
-
-    &__message-container {
-        height: 2rem;
-        padding: 0rem 0rem 0rem 1.6rem;
-        width: 100%;
-
-        &--maxchar {
-            float: right;
-        }
-
-        &--msg {
-            float: left;
-            text-align: left;
-        }
-    }
-}
