--- packages/account-v2/src/components/base/WalletTextField/index.ts
@@ -1,1 +1,0 @@
-export { default as WalletTextField } from './WalletTextField';
