--- packages/account-v2/src/components/base/types.ts
@@ -1,1 +1,0 @@
-export type TGenericSizes = '2xl' | '2xs' | '3xl' | '3xs' | '4xl' | '5xl' | '6xl' | 'lg' | 'md' | 'sm' | 'xl' | 'xs';
